Move menu to gui_base class
